feat: Integrate AgentInfoRow component into AgentsListRow for improved agent display
